% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_rules.R
\name{conversion_rules}
\alias{conversion_rules}
\alias{cache_conversion_rules}
\alias{nuts1}
\alias{nuts2}
\alias{nuts3}
\alias{iso1}
\alias{iso3}
\alias{iso5}
\title{Convert between different region codes}
\usage{
## set up the cache

cache_conversion_rules()

## get the data

nuts1()
nuts2()
nuts3()

iso1()
iso3()
iso5()
}
\description{
Several dataframes containing identifiers for regions of austria.
Every dataset contains three three columns
\itemize{
\item \strong{\code{id}}: The identifier of the region
\item \strong{\code{parent}}: The identifier of the parent region
\item \strong{\code{name}}: The german region name
}

The conversion rules are based on datasets from \url{https://data.statistik.gv.at}.
See the Datasets section below for more details.
}
\section{Datasets from \url{https://data.statistik.gv.at}}{
\itemize{
  \item \strong{\code{OGD_vgrrgr104_RGR104_1}} Gross regional product by ESA 2010 (NUTS2+NUTS3) \href{https://data.statistik.gv.at/web/meta.jsp?dataset=OGD_vgrrgr104_RGR104_1}{metadata} \href{https://data.statistik.gv.at/ogd/json?dataset=OGD_vgrrgr104_RGR104_1}{json}
  \item \strong{\code{OGD_f0743_VZ_HIS_GEM_3}} Population census data since 1869 for communes \href{https://data.statistik.gv.at/web/meta.jsp?dataset=OGD_f0743_VZ_HIS_GEM_3}{metadata} \href{https://data.statistik.gv.at/ogd/json?dataset=OGD_f0743_VZ_HIS_GEM_3}{json}
  \item \strong{\code{OGD_f0743_VZ_HIS_GEM_4}} Population census data since 1869 for Districts and Provinces \href{https://data.statistik.gv.at/web/meta.jsp?dataset=OGD_f0743_VZ_HIS_GEM_4}{metadata} \href{https://data.statistik.gv.at/ogd/json?dataset=OGD_f0743_VZ_HIS_GEM_4}{json}
}
}

