% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{nuts_map}
\alias{nuts_map}
\title{Create a leaflet map}
\usage{
nuts_map(x, join_by = NULL, color = NULL, cutoff = 0.9, overlay_nuts2 = TRUE)
}
\arguments{
\item{x}{a data frame}

\item{join_by}{a column name. This column should contain region codes}

\item{color}{a column name which contains numeric values for coloring}

\item{cutoff}{percentile to cutoff the color scale. Defaults to 90\%}

\item{overlay_nuts2}{show nuts2 boundaries in the map?}
}
\description{
Creates a leaflet map from a data frame which contains spatial codes.
}
\examples{
library(sf)

## visualize population

nuts_map(iso3_pop())
nuts_map(iso5_pop())

nuts_map(nuts2_earnings(), color = "employees")

library(dplyr)

merge(iso5_pop(), iso5_migration()) \%>\%
  mutate(migration_rel = in_migration/pop) \%>\%
  nuts_map(color = "migration_rel")

merge(iso3_pop(), iso3_deaths()) \%>\%
  mutate(deaths_rel = deaths/pop) \%>\%
  nuts_map(color = "deaths_rel")
}
