% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{nutsr_cache_dir}
\alias{nutsr_cache_dir}
\title{Caching in nutsr}
\usage{
nutsr_cache_dir(new = NULL)
}
\arguments{
\item{new}{the new cache dir}
}
\value{
path to the cache directory
}
\description{
The function \code{nutsr_cache_dir()} is invoked by internal methods
of nutsr to dicide where to cache the downloaded files.
}
\details{
Use this function or the environment variable \code{NUTSR_CACHE_DIR} to
change the cache from \code{tempdir()/nutsr} to a persistent storage place.
}
