% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_data.R
\name{nutsr_data}
\alias{nutsr_data}
\alias{iso3_pop}
\alias{iso5_pop}
\alias{nuts2_earnings}
\alias{iso5_migration}
\alias{iso3_deaths}
\title{Data for regions}
\usage{
iso3_pop()

iso5_pop()

nuts2_earnings()

iso5_migration()

iso3_deaths()
}
\description{
Example data for the features of the nutsr package
}
\section{Datasets from \url{https://data.statistik.gv.at}}{
\itemize{
  \item \strong{\code{OGD_f0743_VZ_HIS_GEM_4}} Population census data since 1869 for Districts and Provinces \href{https://data.statistik.gv.at/web/meta.jsp?dataset=OGD_f0743_VZ_HIS_GEM_4}{metadata} \href{https://data.statistik.gv.at/ogd/json?dataset=OGD_f0743_VZ_HIS_GEM_4}{json}
  \item \strong{\code{OGD_f0743_VZ_HIS_GEM_3}} Population census data since 1869 for communes \href{https://data.statistik.gv.at/web/meta.jsp?dataset=OGD_f0743_VZ_HIS_GEM_3}{metadata} \href{https://data.statistik.gv.at/ogd/json?dataset=OGD_f0743_VZ_HIS_GEM_3}{json}
  \item \strong{\code{OGD_veste309_Veste309_1}} Structure of Earnings Survey (SES) 2018 Gross hourly earnings in EUR by citizenship, region (NUTS 2) and form of employment \href{https://data.statistik.gv.at/web/meta.jsp?dataset=OGD_veste309_Veste309_1}{metadata} \href{https://data.statistik.gv.at/ogd/json?dataset=OGD_veste309_Veste309_1}{json}
  \item \strong{\code{OGD_bevwan020_AUSSENWAND_202}} International migration since 2015 by age group, commune and nationality \href{https://data.statistik.gv.at/web/meta.jsp?dataset=OGD_bevwan020_AUSSENWAND_202}{metadata} \href{https://data.statistik.gv.at/ogd/json?dataset=OGD_bevwan020_AUSSENWAND_202}{json}
  \item \strong{\code{OGD_rate_kalwobez_GEST_KALWOCHE_STR_BZ_100}} Age-standardized death rate in Austria by calendar week (excl. abroad) from 2020 by provinces and pol. districts \href{https://data.statistik.gv.at/web/meta.jsp?dataset=OGD_rate_kalwobez_GEST_KALWOCHE_STR_BZ_100}{metadata} \href{https://data.statistik.gv.at/ogd/json?dataset=OGD_rate_kalwobez_GEST_KALWOCHE_STR_BZ_100}{json}
}
}

