% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{nutsr_shapes}
\alias{nutsr_shapes}
\alias{cache_region_shapes}
\alias{iso3_shapes}
\alias{iso5_shapes}
\alias{nuts2_shapes}
\title{Spatial vector data for austrian regions}
\source{
\url{https://github.com/ginseng666/GeoJSON-TopoJSON-Austria}
}
\usage{
cache_region_shapes()

iso3_shapes()

iso5_shapes()

nuts2_shapes()
}
\description{
Function that obtain the polygon shapes for austrian regions.
The data for these polygons is sourced from geojson files and read
via \code{\link[sf:st_read]{sf::st_read()}}.
}
