% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_codes.R
\name{region_codes}
\alias{region_codes}
\title{Find region codes ased on names}
\usage{
region_codes(x, prefer = c("nuts", "iso"))
}
\arguments{
\item{x}{a character vector containing region names}

\item{prefer}{in case of nuts2 regions, this parameter determines
if the nuts2 or the iso representation is returned}
}
\description{
Function to convert region names to region codes. The names must
exactly match an entry of the name column in one of the conversion tables
in order to find a matching code.
}
\examples{
region_codes("Burgenland")
region_codes("Burgenland", prefer = "iso")
}
