% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_labels.R
\name{region_labels}
\alias{region_labels}
\title{Label a vector of region codes}
\usage{
region_labels(x)
}
\arguments{
\item{x}{a vector containing region codes}
}
\value{
a vector containing region labels
}
\description{
Convert region codes to region labels via a lookup in the
region datasets.
}
\examples{
region_labels(c('AT1', 'AT11', 'AT111', '10101'))
}
