% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_levels.R
\name{region_levels}
\alias{region_levels}
\title{Associate regions with an aggregation level}
\usage{
region_levels(x)
}
\arguments{
\item{x}{a character vector containing region codes}
}
\description{
Given a vector of region codes, this function returns the corresponding
region level (nuts1, nuts2, nuts2, iso1, iso3 or iso5) for each entry.
}
\examples{
region_levels(c('AT1', 'AT11', 'AT111', '1', '101', '10101'))
}
