% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_parents.R
\name{region_parents}
\alias{region_parents}
\title{Get parent region}
\usage{
region_parents(x)
}
\arguments{
\item{x}{a vector containing region codes}
}
\description{
Get parent region
}
\examples{
region_parents(c('AT1', 'AT11', 'AT111', '10101'))
}
