% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{sc_roxygen}
\alias{sc_roxygen}
\alias{roxy_tag_rd.roxy_tag_STATcubeR}
\alias{roxy_tag_parse.roxy_tag_STATcubeR}
\alias{format.rd_section_STATcubeR}
\title{The \verb{@STATcubeR} roxygen tag}
\description{
This tag allows you to cite STATcubeR datasets in packages.
It Can be used by placing \verb{@STATcubeR} followed by an \code{id} of an
open data dataset into roxygen comment blocks.\preformatted{#' My Function
#' @STATcubeR OGD_f0743_VZ_HIS_GEM_4
#' @STATcubeR OGD_f0743_VZ_HIS_GEM_3
my_function() \{
  # ...
\}
}

This will fetch metadata for the datasets and render the
resuts into the documentation pages. The output of this example is shown at
the end of this help page.
}
\details{
It is recommended to set up persistent caching with
\code{\link[STATcubeR:od_resource]{STATcubeR::od_cache_dir()}} when using this function.
It is planned to also support datasets from the STATcube REST API in the
future with something like the following syntax\preformatted{#' @STATcubeR sc_table 'path/to/request.json'
#' @STATcubeR sc_example <example_id>``
}
}
\section{Datasets from \url{https://data.statistik.gv.at}}{
\itemize{
  \item \strong{\code{OGD_f0743_VZ_HIS_GEM_4}} Population census data since 1869 for Districts and Provinces \href{https://data.statistik.gv.at/web/meta.jsp?dataset=OGD_f0743_VZ_HIS_GEM_4}{metadata} \href{https://data.statistik.gv.at/ogd/json?dataset=OGD_f0743_VZ_HIS_GEM_4}{json}
  \item \strong{\code{OGD_f0743_VZ_HIS_GEM_3}} Population census data since 1869 for communes \href{https://data.statistik.gv.at/web/meta.jsp?dataset=OGD_f0743_VZ_HIS_GEM_3}{metadata} \href{https://data.statistik.gv.at/ogd/json?dataset=OGD_f0743_VZ_HIS_GEM_3}{json}
}
}

